Change over use of implicit-width types (unsigned int, etc.) with explict types from stdint.h. Reduce modem ring buffer size, alter network.c so that the ring buffer is flushed when it is completely full while processing the uIP buffer contents, rather than every 64 bytes for speed.

+++ trunk/Lib/PPP.c
@@ -45,9 +45,9 @@ PPP_States_t PPPState;											// PPP negotiation states
 // The main loop, login script, PPP state machine
 void PPP_ManagePPPNegotiation(void)
 {
-	int16_t c;													// Received serial character
-	bool ExitFlag = 0;
-	int CharCount = 0;
+	int16_t  c;													// Received serial character
+	bool     ExitFlag = 0;
+	uint16_t CharCount = 0;
 
 	// This will kick-start each phase of the PPP negotiations (LCP, PAP, IPCP)
 	PPP_MakeInitialPacket();
